Cleanup + documentation
