+++ src/views/Skeleton/Skeleton.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import Navbar from '../Navbar';
 import Footer from '../Footer';
-import { pages } from '../../routes';
 
 import './Skeleton.scss';
 import 'scss/General.scss';
@@ -10,9 +9,9 @@ export default class Skeleton extends Component {
    render() {
       return (
          <div>
-            <Navbar items={pages}/>
+            <Navbar />
             {this.props.children}
-            <Footer internalLinks={pages}/>
+            <Footer />
          </div>
       );
    }
